Fix docs on using import
The given example doesn't work on the latest version of webpack (the one I'm running, cannot confirm for others). `addDependencyTo` expects and object with the function `addDependency`, in this case the loader context (as indicated here https://webpack.github.io/docs/loaders.html#adddependency) is actually the one pointing to that function and hence `this` is the object you want. In my case `webpack.addDependency` as `undefined`. Making the suggested change fixed it. Hope it helps and thank you for your contribution to open source :)